Anchor etcd v3 API client under rabbit_sup
We have to start it before the application/plugin is started,
so it cannot be a part of the plugin's supervision tree.
We don't want to simply link it to the calling process, however,
since it's one of the "main" core RabbitMQ server processes.

So rabbit_sup is a sensible compromise.