List existing cluster members via etcd v3 API
Note that we ignore all version difference and assume that
if some keys haven't yet expired, they should be considered.
We then simply filter out duplicates from the returned result
set.

This also means that the key prefix setting is not worth keeping
the way it is: doing a range query with user-provided keys is harder
and error prone, compared to a predictable and constant range query
with a fixed prefix such as "rabbitmq".